Move data structs out of config into data.go
Also move methods to be aligned with the DataStore interface definition order
